feat: simplify issue reporter logic
