+++ apps/next-app/pages/modules/[questionBank]/tests/[testId]/exam.page.tsx
@@ -14,7 +14,6 @@ import {
   TestNavigation,
   TestQuestion,
 } from "@cf/next/tests";
-import { UserBugReport } from "@cf/next/user";
 import { useDisclose } from "@cf/react/components";
 import { ThemeOverrideColorScheme } from "@cf/react/theme";
 import { ssrHandler } from "@cf/trpc/server";
@@ -83,7 +82,6 @@ export const Page: NextPage<Props> = ({ testId, questionBank }) => {
           </DialogContent>
         </ModalDialog>
       </Modal>
-      <UserBugReport />
     </>
   );
 };
