+++ libs/next/user/src/hooks/use-bug-report/use-bug-report.ts
@@ -6,17 +6,13 @@ type DebugCallbacksMap = Record<string, DebugCallback | undefined>;
 
 export const useBugReportStore = create<{
   isOpen: boolean;
-  isAvailable: boolean;
   debugDataCallbacks: DebugCallbacksMap;
   setIsOpen: (isOpen: boolean) => void;
-  setIsAvailable: (isAvailable: boolean) => void;
   setDebugCallback: (key: string, value?: DebugCallback) => void;
 }>((set, get) => ({
   isOpen: false,
-  isAvailable: false,
   debugDataCallbacks: {},
   setIsOpen: (isOpen) => set({ isOpen }),
-  setIsAvailable: (isAvailable) => set({ isAvailable }),
   setDebugCallback: (key, value) => {
     const debugDataCallbacks = { ...get().debugDataCallbacks, [key]: value };
     return set({ debugDataCallbacks });
@@ -25,9 +21,8 @@ export const useBugReportStore = create<{
 
 export const useBugReportDisclose = () => {
   const setIsOpen = useBugReportStore((b) => b.setIsOpen);
-  const isAvailable = useBugReportStore((b) => b.isAvailable);
   const open = useCallback(() => setIsOpen(true), [setIsOpen]);
-  return { isAvailable, open };
+  return { open };
 };
 
 export const useBugReportDebugData = (
