+++ libs/providers/github/src/functions/create-new-issue.ts
@@ -1,3 +1,4 @@
+import { default as YAML } from "yaml";
 import { originOwner, originRepo } from "../common/env";
 import type { NewIssue } from "@cf/core/github";
 import type { Octokit } from "octokit";
@@ -10,13 +11,14 @@ export const createNewIssue = async (octokit: Octokit, newIssue: NewIssue) => {
     repo: originRepo,
     title: `[App Bug Report] ${title}`,
     body: [
-      `## Data\n`,
-      `**href** : ${href}\n`,
-      "```",
-      JSON.stringify(debugData, null, 2),
-      "```",
-      `\n\n---\n\n## Description\n`,
+      "## Description\n",
       description,
+      "\n---\n",
+      "## Data\n",
+      `\n**href** : ${href}`,
+      "```yaml",
+      YAML.stringify(debugData),
+      "```",
     ].join("\n"),
   });
